refactor: optimize tag scroller shuffle to only run on mobile view
