+++ src/components/hero.tsx
@@ -1,9 +1,10 @@
 'use client'
 
 import Link from 'next/link'
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 
 import config from '@/config'
+import { useShuffledTagsOnMobile } from '@/hooks/use-shuffle'
 
 import { Button } from './ui/button'
 import GravityTags from './gravity-tags'
@@ -31,50 +32,8 @@ const tags = [
   { label: '+ more', color: 'bg-[#C3ECF6]' },
 ]
 
-interface Tag {
-  label: string
-  color: string
-}
-
 const Hero = () => {
-  const [shuffledTags, setShuffledTags] = useState(tags)
-
-  const shuffle = (tags: Tag[]): Tag[] => {
-    const arr: Tag[] = [...tags]
-
-    // Shuffle array (Fisher–Yates)
-    for (let i = arr.length - 1; i > 0; i--) {
-      const j = Math.floor(Math.random() * (i + 1))
-
-      ;[arr[i], arr[j]] = [arr[j], arr[i]]
-    }
-
-    // Avoid adjacent duplicates (label or color)
-    for (let i = 0; i < arr.length - 1; i++) {
-      if (
-        arr[i].label === arr[i + 1].label ||
-        arr[i].color === arr[i + 1].color
-      ) {
-        const swapIndex = arr.findIndex(
-          (t: Tag, idx: number) =>
-            idx > i + 1 &&
-            t.label !== arr[i].label &&
-            t.color !== arr[i].color &&
-            arr[i + 1].label !== arr[idx - 1]?.label &&
-            arr[i + 1].color !== arr[idx - 1]?.color,
-        )
-
-        if (swapIndex > -1)
-          [arr[i + 1], arr[swapIndex]] = [arr[swapIndex], arr[i + 1]]
-      }
-    }
-
-    return arr
-  }
-
-  useEffect(() => {
-    setShuffledTags(shuffle(tags))
-  }, [tags])
+  const shuffledTags = useShuffledTagsOnMobile(tags)
 
   return (
     <div className="bg-[#FCF4F4] md:min-h-full pb-12 md:pb-32 relative overflow-hidden">
