+++ src/hooks/use-mobile.ts
@@ -1,0 +1,16 @@
+import { useEffect, useState } from 'react'
+
+export const useIsMobile = (breakpoint = 768) => {
+  const [isMobile, setIsMobile] = useState(false)
+
+  useEffect(() => {
+    const check = () => setIsMobile(window.innerWidth < breakpoint)
+
+    check() // run once
+    window.addEventListener('resize', check)
+
+    return () => window.removeEventListener('resize', check)
+  }, [breakpoint])
+
+  return isMobile
+}
